Mejora de estructura de features, componentes
