+++ src/features/inventory/components/form/FormField.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import type { LucideIcon } from "lucide-react";
+
+interface FormFieldProps {
+  label: string;
+  error?: string;
+  required?: boolean;
+  icon?: LucideIcon;
+  children: React.ReactNode;
+}
+
+const FormField: React.FC<FormFieldProps> = ({ 
+  label, 
+  error, 
+  required = false, 
+  icon: Icon,
+  children 
+}) => {
+  return (
+    <div>
+      <label className="block text-sm font-medium text-gray-700 mb-2">
+        {Icon && <Icon className="w-4 h-4 inline mr-1" />}
+        {label} {required && "*"}
+      </label>
+      {children}
+      {error && <p className="text-red-500 text-xs mt-1">{error}</p>}
+    </div>
+  );
+};
+
+export default FormField;
