+++ src/features/inventory/components/inventory/InventoryView.tsx
@@ -1,0 +1,70 @@
+import React, { useState, useEffect } from "react";
+import { useInventory } from "../../hooks/useInventory";
+import type { NewProduct } from "../../types";
+import ProductTable from "./ProductTable";
+import ProductForm from "./ProductForm";
+import AlertsPanel from "./AlertsPanel";
+import StatsCards from "./StatsCards";
+
+const InventoryView: React.FC = () => {
+  const { 
+    products, 
+    addProduct, 
+    deleteProduct, 
+    updateProductStates, 
+    getStats, 
+    getExpiringProducts,
+    getLowStockProducts 
+  } = useInventory();
+  
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  // Actualizar estados de productos cada vez que se carga el componente
+  useEffect(() => {
+    updateProductStates();
+  }, []);
+
+  const handleAddProduct = (product: NewProduct) => {
+    addProduct(product);
+    setIsModalOpen(false);
+  };
+
+  const stats = getStats();
+  const expiringProducts = getExpiringProducts();
+  const lowStockProducts = getLowStockProducts();
+
+  return (
+    <div className="min-h-screen bg-gray-50">
+      <div className="mt-6" />
+
+      {/* Componente de Alertas */}
+      <AlertsPanel 
+        expiringProducts={expiringProducts}
+        lowStockProducts={lowStockProducts}
+      />
+
+      {/* Contenido principal */}
+      <div className="p-6 pt-4">
+        {/* Componente de Estadísticas */}
+        <StatsCards stats={stats} />
+
+        {/* Tabla de productos */}
+        <ProductTable 
+          products={products} 
+          onDelete={deleteProduct}
+          onAddProduct={() => setIsModalOpen(true)}
+        />
+      </div>
+
+      {/* Modal de formulario */}
+      {isModalOpen && (
+        <ProductForm
+          onSubmit={handleAddProduct}
+          onClose={() => setIsModalOpen(false)}
+        />
+      )}
+    </div>
+  );
+};
+
+export default InventoryView;
