+++ src/pages/InventoryPage/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import InventoryView from "../../features/inventory/components/InventoryView";
+import InventoryView from "../../features/inventory/components/inventory/InventoryView";
 
 const InventoryPage: React.FC = () => {
   return <InventoryView />;
